+++ packages/clients/diplan/vite.config.js
@@ -1,0 +1,18 @@
+import { getClientConfig } from '../../../viteConfigs'
+
+export default getClientConfig({
+  root: 'example',
+  build: {
+    lib: {
+      name: 'MapClient',
+      entry: '../src/polar-client.ts',
+      fileName: () => 'polar-client.js',
+    },
+    rollupOptions: {
+      output: {
+        assetFileNames: (assetInfo) =>
+          assetInfo.name === 'style.css' ? 'polar-client.css' : assetInfo.name,
+      },
+    },
+  },
+})